requirements.txt
# 
azure-mgmt-storage
azure-storage-blob
azure-mgmt-resource
azure-keyvault-secrets
azure-identity
# Manage Resource
azure-mgmt-resource>=2.1.0
azure-mgmt-storage>=3.0.0
haikunator